fix(banking): stop leaking raw errors & PII from sync path (re-audit #19/#20)
From AUDIT_2026-06-12_DEEP_REAUDIT.md, banking observability hardening:

- #19 raw error leakage: the 13 per-handler copies of createErrorResponse each
  accepted a `details` arg that was spread into the JSON response, exposing raw
  Supabase/driver errors (schema, constraint names, ids) to callers. Replaced
  them with a single shared api/_lib/http-error.ts helper that logs `details`
  server-side and returns only { error, code }. sync-transactions also now
  returns a generic "Transaction sync failed" instead of the raw error.message.

- #20 PII in logs: removed debug console.log lines that dumped bank
  external_account_ids and a full TrueLayer transaction sample (merchant,
  amount, description) to server logs. Sync counts remain in sync_history for
  observability without the PII.

Net effect: even handlers that still pass an error object as `details` no longer
leak it — the shared helper logs, never serialises, it. Removes ~13 duplicated
helper definitions.

Verified: typecheck:strict, lint, build all clean.
