+++ api/banking/dead-letter-admin-audit.ts
@@ -2,12 +2,12 @@ import { Buffer } from 'node:buffer';
 import type { VercelRequest, VercelResponse } from '@vercel/node';
 import type {
   DeadLetterAdminAuditResponse,
-  DeadLetterAdminAuditRow,
-  ErrorResponse
+  DeadLetterAdminAuditRow
 } from '../../src/types/banking-api.js';
 import { AuthError, requireAuth } from '../_lib/auth.js';
 import { requireBankingOpsAdmin } from '../_lib/banking-ops.js';
 import { setCorsHeaders } from '../_lib/cors.js';
+import { createErrorResponse } from '../_lib/http-error.js';
 import { getServiceRoleSupabase } from '../_lib/supabase.js';
 
 const DEFAULT_LIMIT = 25;
@@ -30,20 +30,6 @@ interface DeadLetterAuditDbRow {
   created_at: string;
   completed_at: string | null;
 }
-
-const createErrorResponse = (
-  res: VercelResponse,
-  status: number,
-  error: string,
-  code: string,
-  details?: unknown
-) => {
-  const payload: ErrorResponse = { error, code };
-  if (details !== undefined) {
-    payload.details = details;
-  }
-  return res.status(status).json(payload);
-};
 
 const relationMissing = (error: unknown): boolean => {
   if (!error || typeof error !== 'object') {
