+++ api/banking/dead-letter-admin.ts
@@ -2,8 +2,7 @@ import type { VercelRequest, VercelResponse } from '@vercel/node';
 import type {
   DeadLetterAdminListResponse,
   DeadLetterAdminResetRequest,
-  DeadLetterAdminResetResponse,
-  ErrorResponse
+  DeadLetterAdminResetResponse
 } from '../../src/types/banking-api.js';
 import { AuthError, requireAuth } from '../_lib/auth.js';
 import {
@@ -12,26 +11,13 @@ import {
   requireBankingOpsAdmin
 } from '../_lib/banking-ops.js';
 import { setCorsHeaders } from '../_lib/cors.js';
+import { createErrorResponse } from '../_lib/http-error.js';
 import { getServiceRoleSupabase } from '../_lib/supabase.js';
 
 const RESET_ALL_CONFIRMATION = 'RESET_ALL_DEAD_LETTERED';
 const DEFAULT_PREVIEW_LIMIT = 25;
 const MAX_PREVIEW_LIMIT = 200;
 const MAX_RESET_LIMIT = 500;
-
-const createErrorResponse = (
-  res: VercelResponse,
-  status: number,
-  error: string,
-  code: string,
-  details?: unknown
-) => {
-  const payload: ErrorResponse = { error, code };
-  if (details !== undefined) {
-    payload.details = details;
-  }
-  return res.status(status).json(payload);
-};
 
 const relationMissing = (error: unknown): boolean => {
   if (!error || typeof error !== 'object') {
