+++ api/banking/ops-alert-stats.ts
@@ -1,5 +1,5 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node';
-import type { ErrorResponse, OpsAlertStatsResponse } from '../../src/types/banking-api.js';
+import type { OpsAlertStatsResponse } from '../../src/types/banking-api.js';
 import { AuthError, requireAuth } from '../_lib/auth.js';
 import {
   getOpsAlertNotifyEvery,
@@ -7,6 +7,7 @@ import {
   requireBankingOpsAdmin
 } from '../_lib/banking-ops.js';
 import { setCorsHeaders } from '../_lib/cors.js';
+import { createErrorResponse } from '../_lib/http-error.js';
 import { getServiceRoleSupabase } from '../_lib/supabase.js';
 
 const DEFAULT_LIMIT = 50;
@@ -19,20 +20,6 @@ interface OpsAlertCounterRow {
   suppressed_count: number;
   updated_at: string | null;
 }
-
-const createErrorResponse = (
-  res: VercelResponse,
-  status: number,
-  error: string,
-  code: string,
-  details?: unknown
-) => {
-  const payload: ErrorResponse = { error, code };
-  if (details !== undefined) {
-    payload.details = details;
-  }
-  return res.status(status).json(payload);
-};
 
 const parsePositiveInt = (
   value: string | undefined,
